% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{find_manual_errors}
\alias{find_manual_errors}
\title{Identify object labels that are not present in image masks.}
\usage{
find_manual_errors(experiment, objects_data_frame)
}
\arguments{
\item{experiment}{A list that has an \code{experiment} name and \code{id} as a Synapse ID for a file with the survival data in it.}

\item{objects_data_frame}{A data frame of potential objects.}
}
\value{
A data frame summary of the objects not present.
}
\description{
These are potentially manual mistakes during curation.
}
