% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{get_censored_wells}
\alias{get_censored_wells}
\title{Get censored wells from the censored well table.}
\usage{
get_censored_wells(id = "syn11709601", ...)
}
\arguments{
\item{id}{Synapse ID of a censored well table.}

\item{...}{Additional parameters to synTableQuery}
}
\value{
A data frame of censored well information.
}
\description{
Get censored wells from the censored well table.
}
